back-end/src/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
boto3==1.26.49
botocore==1.29.104
braintree==4.19.0
certifi==2021.10.8
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.0.8
click==8.1.3
croniter==1.3.4
cryptography==39.0.0
cypress-django==0.1.0
dataclasses==0.6
ddtrace==0.56.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.2
django-appconf==1.0.5
django-bootstrap4==22.1
django-ckeditor==6.2.0
django-compressor==4.3.1
django-cors-headers==3.14.0
django-crum==0.7.9
django-debug-toolbar==3.4.0
django-embed-video==1.4.0
django-environ==0.8.1
django-extensions==3.1.5
django-htmx==1.9.0
django-js-asset==2.0.0
django-libsass==0.9
django-redis==5.0.0
django-rq==2.5.1
djangorestframework==3.14.0
ERPpeek==1.7.1
et-xmlfile==1.1.0
fakeredis==1.8
filelock==3.6.0
gunicorn==20.0.2
html2text==2020.1.16
identify==2.5.0
idna==3.3
jmespath==1.0.1
meilisearch==0.19.1
mysql-connector-python==8.0.28
nodeenv==1.6.0
openpyxl==3.0.9
packaging==21.3
Pillow==8.4.0
platformdirs==2.5.2
pre-commit==2.18.1
protobuf==3.20.1
psycopg2-binary==2.9.3
py-spy==0.3.11
pyclean==2.2.0
pycparser==2.21
pyparsing==3.0.8
pysodium==0.7.11
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.2
pytz==2021.3
PyYAML==6.0
rcssmin==1.1.1
redis==4.2.2
requests==2.26.0
rjsmin==1.2.1
rq==1.10.1
rq-scheduler==0.11.0
s3transfer==0.6.0
sentry-sdk==1.9.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tenacity==8.0.1
toml==0.10.2
typing_extensions==4.5.0
urllib3==1.26.7
virtualenv==20.14.1
Werkzeug==2.1.2
wrapt==1.14.1
